% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{manhattan_plot
Manhattan plot showing on/off-targets detected organized by chromosomal
position with bar height representing read count (number of breaks).}
\usage{
manhattan_plot(
  x,
  bsgenome,
  standard_chromosomes = TRUE,
  cutsite_breaks_only = TRUE,
  subtract_nontarget = FALSE,
  log_signal = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis. Possibly filtered (eg. qval < .01).}

\item{bsgenome}{character, bsgenome to use (eg. BSgenome.Hsapiens.UCSC.hg38).}

\item{standard_chromosomes}{logical, constrain the analysis to standard.
chromosomes only. Default: TRUE.}

\item{cutsite_breaks_only}{display breaks of cutsite only, instead of using
breaks accumulated in the complete protospacer. Default: TRUE}

\item{subtract_nontarget}{subtract breaks of non-target library.
Default: FALSE}

\item{log_signal}{display log2 transformed signal. Default: FALSE.}

\item{verbose}{logical, keep informing about every step.}
}
\value{
A ggplot object.
}
\description{
manhattan_plot
Manhattan plot showing on/off-targets detected organized by chromosomal
position with bar height representing read count (number of breaks).
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )
}
data(breakinspectoR_examples)

manhattan_plot(
  offtargets,
  bsgenome ="BSgenome.Hsapiens.UCSC.hg38"
)
}
