% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakinspector.R
\name{msg}
\alias{msg}
\title{msg}
\usage{
msg(verbose, ...)
}
\arguments{
\item{verbose}{logical, keep informing about every step}

\item{...}{other parameters sent to `message()`}
}
\value{
nothing
}
\description{
Write messages to the console
}
