% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakinspector-package.R
\docType{data}
\name{offtargets.scission_profile}
\alias{offtargets.scission_profile}
\title{Scission profile of the offtargets detected within the BreakTag test data
included in this package}
\format{
A GRanges object with the coordinates of the cutsite, the number of
breaks on and around the cutsite for the target and non-target libraries,
and a p- and q-values, as a result from callilng the breakinspectoR analysis
with `extdata/vegfa.chr6.bed.gz` and 'extdata/nontarget.chr6.bed.gz'.
}
\usage{
data(breakinspectoR_examples)
}
\description{
A GRanges object with the coordinates of the cutsite, the number of
breaks on and around the cutsite for the target and non-target libraries,
and a p- and q-values.
}
\details{
It is the result of running `scission_profile_analysis()` with the example
datasets included in this package:
`extdata/vegfa.chr6.bed.gz` and 'extdata/nontarget.chr6.bed.gz'
}
\keyword{datasets}
