% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_genomic_position}
\alias{plot_genomic_position}
\title{plot_genomic_position
Plot the genomic position of the breaks, identifying detected
on-/off-targets.}
\usage{
plot_genomic_position(
  x,
  target,
  bsgenome,
  standard_chromosomes = TRUE,
  min_breaks = 2,
  log_signal = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis. Possibly filtered (eg. qval < .01).}

\item{target}{GRanges object with coordinates of breaks in the targeted
library, or a character with the path to a bed file.}

\item{bsgenome}{character, bsgenome to use (eg. BSgenome.Hsapiens.UCSC.hg38).}

\item{standard_chromosomes}{logical, constrain the analysis to standard
chromosomes only. Default: TRUE.}

\item{min_breaks}{minimum number of breaks in the cutsite to be considered
for the analysis. Default: 2.}

\item{log_signal}{display log10 transformed signal. Default: FALSE.}

\item{verbose}{logical, keep informing about every step.}
}
\value{
A ggplot object.
}
\description{
plot_genomic_position
Plot the genomic position of the breaks, identifying detected
on-/off-targets.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )
}
data(breakinspectoR_examples)

plot_genomic_position(
  offtargets,
  target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
  bsgenome ="BSgenome.Hsapiens.UCSC.hg38"
)
}
