% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_guide_fidelity}
\alias{plot_guide_fidelity}
\title{plot_guide_fidelity}
\usage{
plot_guide_fidelity(x, guide, pam)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{guide}{character with the guide sequence. Only used for drawing, the
number of mismatches is already present in `x`.}

\item{pam}{character with PAM sequence. Only used for drawing, the number of
mismatches is already present in `x`.}
}
\value{
Nothing.
}
\description{
Show, for each position of the guide, the fidelity of the offtarget sequence
as the percentage of matching bases.
}
\details{
Currently this function has the limitation that does only *fixed* matching,
expecting the offtarget sequence and the guide to show the same character.
See ?`Biostrings::lowlevel-matching` for more information.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )

  plot_guide_fidelity(offtargets, guide="GACCCCCTCCACCCCGCCTC", pam="NGG")
}
}
