% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_mismatch_freq}
\alias{plot_mismatch_freq}
\title{plot_mismatch_freq
Plot the frequency of mismatches per offtarget.}
\usage{
plot_mismatch_freq(x, bins = 1 + diff(range(x$mismatches)))
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{bins}{number of bins tu use in the histogram. Defaults to the maximum
number of mismatches + 1.}
}
\value{
A ggplot object.
}
\description{
plot_mismatch_freq
Plot the frequency of mismatches per offtarget.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )

  plot_mismatch_freq(offtargets)
}
}
