% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_offtargets_by_pam}
\alias{plot_offtargets_by_pam}
\title{plot_offtargets_by_pam
Plot the count of offtargets by the different PAMs}
\usage{
plot_offtargets_by_pam(x, fraction = TRUE)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{fraction}{logical indicating whether the plot should show absolute
absolute numbers or the fraction.}
}
\value{
A ggplot object.
}
\description{
plot_offtargets_by_pam
Plot the count of offtargets by the different PAMs
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )

  plot_offtargets_by_pam(offtargets, fraction=TRUE)
}
}
