% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_overhang_size}
\alias{plot_overhang_size}
\title{plot_overhang_size
Plots the proportion of blunt and staggered reads by size.}
\usage{
plot_overhang_size(x)
}
\arguments{
\item{x}{A named list of GRanges object containing the results of a
scission_profile_analysis.}
}
\value{
A ggplot object.
}
\description{
plot_overhang_size
Plots the proportion of blunt and staggered reads by size.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )

  offtargets.scission_profile <- scission_profile_analysis(
    x        =offtargets,
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38")
}
data(breakinspectoR_examples)

# simulate 2 different experiments by picking 25 random offtargets
exp1 <- sample(offtargets.scission_profile, 25)
exp2 <- sample(offtargets.scission_profile, 25)
plot_overhang_size(list(a=exp1, b=exp2))
}
