% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_position_cutsite}
\alias{plot_position_cutsite}
\title{plot_position_cutsite
Plot the frequency of the position of the cutsite relative to the guide.}
\usage{
plot_position_cutsite(x, guide, pam)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{guide}{character with the guide sequence. Only used for drawing.}

\item{pam}{character with PAM sequence. Only used for drawing.}
}
\value{
A ggplot object.
}
\description{
plot_position_cutsite
Plot the frequency of the position of the cutsite relative to the guide.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )
}
data(breakinspectoR_examples)

plot_position_cutsite(offtargets, guide="GACCCCCTCCACCCCGCCTC", pam="NGG")
}
