% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_sequence_composition}
\alias{plot_sequence_composition}
\title{plot_sequence_composition
Sequence composition of off-target sites.}
\usage{
plot_sequence_composition(x, guide, pam)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{guide}{character with the guide sequence. Only used for drawing, the
number of mismatches is already present in `x`.}

\item{pam}{character with PAM sequence. Only used for drawing, the number of
mismatches is already present in `x`.}
}
\value{
A ggplot object.
}
\description{
plot_sequence_composition
Sequence composition of off-target sites.
}
\examples{
\dontrun{
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    cutsiteFromPAM=3
  )
}
data(breakinspectoR_examples)

plot_sequence_composition(offtargets, guide="GACCCCCTCCACCCCGCCTC", pam="NGG")
}
