% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakinspector.R
\name{read_targets}
\alias{read_targets}
\title{read_targets}
\usage{
read_targets(x, genome, standard_chromosomes, strandless)
}
\arguments{
\item{x}{GRanges object with coordinates of breaks or the path to a bed file.}

\item{genome}{a BSgenome object.}

\item{standard_chromosomes}{logical, constrain the analysis to standard
chromosomes only (defaults to TRUE).}

\item{strandless}{collapse reads from multiple strands.}
}
\value{
A GRanges object, unstranded and with a `score` column
}
\description{
Read in a bed file with positions of breaks
}
\examples{
breaks <- read_targets(x=system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
                       genome="BSgenome.Hsapiens.UCSC.hg38",
                       standard_chromosomes=TRUE,
                       strandless=FALSE)
}
