% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakinspector.R
\name{reduceOT}
\alias{reduceOT}
\title{reduceOT}
\usage{
reduceOT(
  x,
  fdr = 0.01,
  qval = 0.01,
  mismatches = 7,
  standard_chromosomes = TRUE,
  cores = getOption("mc.cores", 2L),
  verbose = TRUE
)
}
\arguments{
\item{x}{A GRanges object containing the results of a breakinspectoR
analysis.}

\item{fdr}{numerical, keep only hits below this empirical FDR threshold.}

\item{qval}{numerical, keep only hits below this q-value threshold.}

\item{mismatches}{numeric, mismatches allowed between the guide and the
genomic sequence.}

\item{standard_chromosomes}{logical, constrain the analysis to standard
chromosomes only (defaults to TRUE).}

\item{cores}{numerical, number of parallel cores to use.}

\item{verbose}{logical, keep informing about every step}
}
\value{
a GRanges object with the filtered OT.
}
\description{
Does a very opinionated filtering of the OT detected by breakinspector.
}
\examples{
## this is needed only for the package to install
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly=TRUE)) {
##
  offtargets <- breakinspectoR(
    target   =system.file("extdata/vegfa.chr6.bed.gz", package="breakinspectoR"),
    nontarget=system.file("extdata/nontarget.chr6.bed.gz", package="breakinspectoR"),
    guide    ="GACCCCCTCCACCCCGCCTC",
    PAM      =c("NGG", "NAG"),
    bsgenome ="BSgenome.Hsapiens.UCSC.hg38",
    eFDR     =FALSE,
    cutsiteFromPAM=3)

  offtargets_filtered <- reduceOT(offtargets)
}

}
