% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_bluntPred.R
\name{shiny_bluntPred}
\alias{shiny_bluntPred}
\title{shiny_bluntPred
Launch an interactive web interface.}
\usage{
shiny_bluntPred(...)
}
\arguments{
\item{...}{other params sent to shiny::runApp().}
}
\value{
Nothing
}
\description{
shiny_bluntPred
Launch an interactive web interface.
}
