% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_BTmotif.R
\name{shiny_BTmotif}
\alias{shiny_BTmotif}
\title{shiny_BTmotif
Launch an interactive web interface.}
\usage{
shiny_BTmotif(...)
}
\arguments{
\item{...}{other params sent to shiny::runApp().}
}
\value{
Nothing
}
\description{
shiny_BTmotif
Launch an interactive web interface.
}
